ui_requirements.txt
gradio==4.13.0
python-dotenv==1.0.0
openai==1.3.0
mem0==0.1.0 